+++ src/data/+mag/TimeSeries.m
@@ -32,6 +32,9 @@ classdef (Abstract) TimeSeries < mag.Data
 
         % RESAMPLE Resample data to the specified frequency.
         resample(this, targetFrequency)
+
+        % DOWNSAMPLE Downsample data to the specified frequency.
+        downsample(this, targetFrequency)
     end
 
     methods (Hidden, Sealed)