+++ resources/ReleaseNotes.md
@@ -2,5 +2,6 @@
 - Customize display of `mag.Science` and `mag.HK` classes
 - Always show custom display of `mag.Instrument` when meta data is available
 - Update comment in app on which export formats do not support start and end dates
+- Move definition of `downsample` method to `mag.TimeSeries` class
 - Fix typos in `resample` methods help
 - Fix issue with plotting a `stackedplot` without y-axis labels